Move android commands to android package
